requirements.txt
pynacl==1.2.*
django>=1.11
